Initial changes for the artist shop section
