+++ src/pages/moderator/cosmetic-store/cosmetics/index.tsx
@@ -49,100 +49,6 @@ import { CosmeticGetById } from '~/types/router';
 
 import { trpc } from '~/utils/trpc';
 
-const cosmeticSampleSizeMap: Record<
-  'sm' | 'md' | 'lg',
-  { badgeSize: number; textSize: MantineSize; avatarSize: MantineSize }
-> = {
-  sm: { badgeSize: 50, textSize: 'sm', avatarSize: 'md' },
-  md: { badgeSize: 80, textSize: 'md', avatarSize: 'xl' },
-  lg: { badgeSize: 120, textSize: 'lg', avatarSize: 'xl' },
-};
-
-export const CosmeticSample = ({
-  cosmetic,
-  size = 'sm',
-}: {
-  cosmetic: Pick<CosmeticGetById, 'id' | 'data' | 'type' | 'name'>;
-  size?: 'sm' | 'md' | 'lg';
-}) => {
-  const values = cosmeticSampleSizeMap[size];
-
-  switch (cosmetic.type) {
-    case CosmeticType.Badge:
-    case CosmeticType.ProfileDecoration:
-      const decorationData = cosmetic.data as BadgeCosmetic['data'];
-      if (!decorationData.url) return null;
-
-      return (
-        <Box w={values.badgeSize}>
-          <EdgeMedia src={decorationData.url} alt={cosmetic.name} />
-        </Box>
-      );
-    case CosmeticType.ContentDecoration:
-      const contentDecorationData = cosmetic.data as ContentDecorationCosmetic['data'];
-      if (!contentDecorationData.url && !contentDecorationData.cssFrame) return null;
-
-      return (
-        <Box w={values.badgeSize}>
-          <FeedCard
-            aspectRatio="square"
-            frameDecoration={cosmetic as ContentDecorationCosmetic}
-            sx={{ margin: '0 !important' }}
-          >
-            <Box
-              w="100%"
-              h="100%"
-              sx={(theme) => ({
-                backgroundColor:
-                  theme.colorScheme === 'dark' ? theme.colors.dark[7] : theme.colors.gray[1],
-              })}
-            />
-          </FeedCard>
-        </Box>
-      );
-    case CosmeticType.NamePlate:
-      const data = cosmetic.data as NamePlateCosmetic['data'];
-      return (
-        <Text weight="bold" {...data} size={values.textSize}>
-          Sample Text
-        </Text>
-      );
-    case CosmeticType.ProfileBackground:
-      const backgroundData = cosmetic.data as ProfileBackgroundCosmetic['data'];
-      if (!backgroundData.url) return null;
-
-      return (
-        <Box
-          style={{
-            height: values.badgeSize,
-            width: '100%',
-            overflow: 'hidden',
-            borderRadius: 10,
-          }}
-        >
-          <EdgeMedia
-            src={backgroundData.url}
-            alt={cosmetic.name}
-            type={backgroundData.type}
-            anim={true}
-            style={{
-              objectFit: 'cover',
-              // objectPosition: 'right bottom',
-              width: '100%',
-              height: '100%',
-            }}
-            wrapperProps={{
-              style: { height: '100%' },
-            }}
-            contain
-          />
-        </Box>
-      );
-    default:
-      return null;
-  }
-};
-
 export const CosmeticPreview = ({
   cosmetic,
 }: {
